+++ pkg/R/vcovPL.R
@@ -16,7 +16,7 @@ vcovPL <- function(x, cluster = NULL, order.by = NULL,
 }
 
 meatPL <- function(x, cluster = NULL, order.by = NULL,
-  kernel = "Bartlett", lag = "NW1987", bw = NULL, adjust = TRUE, ...) ## adjust/cadjust?
+  kernel = "Bartlett", lag = "NW1987", bw = NULL, adjust = TRUE, aggregate = TRUE, ...) ## adjust/cadjust?
 {
   ## extract estimating functions / aka scores
   if (is.list(x) && !is.null(x$na.action)) class(x$na.action) <- "omit"
@@ -90,10 +90,15 @@ meatPL <- function(x, cluster = NULL, order.by = NULL,
   order.by <- order.by[index]
   cluster <- cluster[index]
 
-  ## aggregate within time periods
-  if(length(unique(order.by)) < n) ef <- apply(ef, 2L, tapply, order.by, sum)
-  nt <- NROW(ef)
-  
+  ## aggregate within time periods?
+  if(aggregate) {
+    if(length(unique(order.by)) < n) ef <- apply(ef, 2L, tapply, order.by, sum)
+    nt <- NROW(ef)
+  } else {
+    order.by <- as.integer(factor(order.by))
+    nt <- order.by[length(order.by)]
+  }  
+
   ## lag/bandwidth selection
     if(is.character(lag)) {
         if(lag == "P2009") lag <- "max" 
@@ -114,12 +119,17 @@ meatPL <- function(x, cluster = NULL, order.by = NULL,
 
   ## set up kernel weights up to maximal number of lags
   weights <- kweights(0L:(nt - 1L)/bw, kernel = kernel)
+  weights <- weights[1L:max(which(abs(weights) > 0))]
 
   rval <- 0.5 * crossprod(ef) * weights[1L]
     
   if(length(weights) > 1L) {
     for (ii in 2L:length(weights)) {
-      rval <- rval + weights[ii] * crossprod(ef[1L:(nt - ii + 1), , drop = FALSE], ef[ii:nt, , drop = FALSE])
+      rval <- rval + weights[ii] * if(aggregate) {
+        crossprod(ef[1L:(nt - ii + 1L), , drop = FALSE], ef[ii:nt, , drop = FALSE]) ## Driscoll & Kraay (1998)
+      } else {
+        crossprod(ef[order.by %in% (1L:(nt - ii + 1L)), , drop = FALSE], ef[order.by %in% (ii:nt), , drop = FALSE]) ## restricting cross-sectional and cross-serial correlation to zero
+      }
     }
  }